+++ src/main/java/com/example/researching/service/DynamicParamsStrategyService.java
@@ -1,10 +1,12 @@
 package com.example.researching.service;
 
+import java.math.BigDecimal;
+
 /**
  * @author Kenny Fang
  * @since 0.0.2
  */
 public interface DynamicParamsStrategyService {
 
-    float calculateFee(String... args);
+    BigDecimal calculateFee(String... args);
 }
